feat: set collisions to engineers.
All collisions can be assigned their own engineers. Fixed the order of
adding collisions to the database, mappers adapted to empty values. Each
engineer's page displays lists of clashes that are assigned to him.
Added HomeController. The total number of engineers and collisions is
displayed. Added functionality for page navigation