feat: optimize SQL requests in DAOs and repositories
This commit includes changes to the DAOs and repositories in order to optimize SQL requests to the database. The modifications were made to minimize the number of SQL requests and improve the performance of the application.

Specifically, the methods in the DAOs and repositories were updated to use more efficient SQL queries and reduce the amount of data being retrieved from the database.

This implementation will lead to faster application response times and improved overall performance.

Signed-off-by: Valentin Mitrofanov <mitrofmep@gmail.com>